fixed norminette except for a few: solution.c: line 68 (line too long); line 127 (function has more than 25 lines). block.c: line 71 (function has more than 25 lines)
