+++ src/fillit.c
@@ -17,9 +17,7 @@ int	main(int argc, char *argv[])
 	char	*filecontent;
 
 	if (argc != 2)
-	{
 		ft_putstr("usage: fillit source_file\n");
-	}
 	filecontent = read_file(argv[1]);
 	if (filecontent == NULL)
 		return (0);
@@ -29,9 +27,7 @@ int	main(int argc, char *argv[])
 		return (0);
 	}
 	if (!final_solution(filecontent))
-	{
 		return (0);
-	}
 	free(filecontent);
 	return (0);
 }
